Ajout des modes de fonctionnement du module feu rouge

+++ lightkit/software/src/io/outputs.hpp
@@ -5,15 +5,15 @@
   * @date   12/08/2019
   */
 
-#include <Arduino.h>
-
 #ifndef IO_OUTPUTS_HPP
 #define IO_OUTPUTS_HPP
+
+#include "io.hpp"
 
 #ifdef MODULE_OUTPUTS
 
 int  outputs_init();
-void output_set(uint8_t i, bool state);
+void output_set(uint32_t i, bool state);
 void output_delayed_set(uint8_t i, bool state, uint32_t delay);
 bool output_get(uint8_t i);
 void output_main(void);
